CodeHax/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
groq==0.10.0
python-dotenv==1.0.0
httpx==0.25.2

# MongoDB
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Additional utilities
python-dateutil==2.8.2
fastapi-mail